requirements.txt
streamlit
pandas
spacy
newspaper3k
nltk
isort
autoflake
autopep8
black